Added the ablity to search for multiple names in the party name search. (#159)
The issue was that the query searched for first name and party name separately.
So if a person searched a Bob Smith, it would look for a Bob Smith as a first name
and Bob Smith as a party name and return no results.

To solve this, I split the incoming search string into separate word and search
for each word individually. So Bob and Smith would be searched in the first_name
and party_name.