Feature/mds 472 (#196)
* Add additional description to swagger documentation.

Provided additional parameter information and description to swagger
documentation that can be accessed through the root url.

* Restructure file to have namespaces in api

Modularized backend to mines, parties, and permits and created namespaces
for each module. This will help with organization and as features
increase there might be a need to break things into its own app or
microservice.