Fix party search to include organization type
There was an issue where party search was working with person type for party, first name and last name
was working properly. Unfortunately the search was not working with organization. The issue was that
the hybrid_property used to make 'name' for Party, was not working for queries, a .expression was used
and sqlalchemy func with concat solved the issue.

There were a couple of code clean up done, as the search was placed into the Party class, and a
couple of test cases were added to test implicit and explicit search.