Sprint2018.9 (#176)
* Feature/mds 457 (#173)

* updated file names

* added popup confirmation to modals, changed directory names under mine

* Update USAGE.md to include a Getting Started section for Windows as well as the command to create test data.

* Update USAGE.md

* Fix the UnitTest label (#177)

* added environment variables (#178)

fixes infinite spinner, adds path prefix to URI.

* Refactor/backend smell test (#179)

* Refactor create_data function to decrease viewing complexity.

Altered the refactoring of creating data to decrease viewing complexity.
There were too many nested for loops and nested loops so I broke up
the giant functions to smaller functions, and added helper functions in
the respective model classes to help create the objects.

* Increase test coverage and include missing test cases.

The purpose of this commit is to increase the test coverage to go
back over 90%. I also added the missing test coverage for
model unit tests and particularly testing for AssertionErrors.

* Feature/mds 446 - Add Clustering (#175)

* Add clustering

* remove className

* update HTTPS url

* fix redux hot reloading

* Add local hosting for fcl

* Add a readme

* rename scss files

* Changed api routes to plurals (#180)

* Changed api routes to plurals
This was a change previously discussed where api routes will
take its plural form and not split into singular and plural.
It will make the api routes more consistent. The frontend constants
file was changed to follow the changes.

* Add Error Mixin in MineLocationResource.

Added error mixin into mine_location resource to keep
error messaging consistent.

* Add permit no as a search filter in mine search. (#181)

Added the permit no in the mine list search filter.

* Add tabnames to url route support (#183)

* Feature/mds 457 (#185)

* Add additional description to swagger documentation. (#182)

* Add additional description to swagger documentation.

Provided additional parameter information and description to swagger
documentation that can be accessed through the root url.

* Add support for path prefixes

* Add search query into mine resource get (#190)

Added search query into mine resource get to return a paginated
list of the search term. The query string used is ?search.

* Added a new view mine manager history (#188)

* Added a new landing page for Unauthorized visitors (#184)

* updated null screen for mine manager history (#193)

* Mds 405/feature/search by permit (#192)

* Mds 405/feature/search by permit (#194)

* Flexible Resource Allocation (#186)

* bugfix/FCLDataLayer

* Feature/mds 450 (#197)

* update .env example

* update test codes

* add waitFor condition

* Update .gitignore

* Feature/mds 472 (#196)

* Add additional description to swagger documentation.

Provided additional parameter information and description to swagger
documentation that can be accessed through the root url.

* Restructure file to have namespaces in api

Modularized backend to mines, parties, and permits and created namespaces
for each module. This will help with organization and as features
increase there might be a need to break things into its own app or
microservice.

* Fix incorrect url constants. (#198)

A 404 has been occuring, because the url is missing a slash,
it was hitting a mispelled endpoint.

To fix this, we added a missing slash to both /parties/managers and
/permits/permittees endpoint.

* Refactor/selectors (#195)

* Backend optimization (#199)

* Update record creation to use threading

* Add database index

* optimize minelist query

* Commit often to reduce race conditions between tests (#200)

* Restructure apiversion over name for postgres DC

* fix postgres resource typo

* Add database dump scripts (#201)

* added database restore and dump scripts

* add makefile commands

* update the restore script

* updated addingParty (#202)

* Add meta tag to index.html (#203)

* Bug/mds 322 (#205)

* Bug/mds 322 (#206)