data tables creation
mds_required_document are reports described in the code, it holds the reporting requirement

mine_file relates the file to the mine, for now there will not be a file, but this object is needed to actually link a mine to the reporting requirement.

required_document_files relates the mine_file to the mds_required_document in a many to many relationship probably a multi-select list (X document is related to Y reporting requirement).