MDS-481 bug/fixed forms from showing initialValues after submission (#220)
* Add Migration and Undo Migration Script

Added migration script for adding a column for major or regional mine,
since it is only one or the other a boolean column of 'major' was
created and will default to false.

* Update mine_detail model and mine resource

Updated mine_detail model and mine resource file to accept
'major' in /mines POST and PUT json payload. Note that 'is None' had
to be used as python 'False' is used as a negative value in boolean
logic.

* Add tests for mine_detail major

Added tests for creating mines with major or regional, and updating
existing mines.

* Fix error expiry_date on models

There were a couple of places in location and tenure where the expiry
date was using 'now' instead of '9999-12-31' as the expiry date.

* cleaned up imports, by adding a config file

* add returned value

* add region description in returned json

* updated tests

* updated region

* updated view to show region

* updated create/edit mine

* Clear form by default and provide an opt-out

While most forms should be reset on submission, the form for updating a
mine record should retain its updated values while the submission is
processing. Given no 'clear' value, a form will continue to clear by
default, not affecting the behaviour of existing forms.