MDS-454 Feature/Support disturbance deletion and add validation (#306)
* Add PUT /mines/mine_types and frontend logic

The frontend now makes separate requests when updating a mine record:
one for the mine_identity/details and one for mine_type. This will be
built-upon with disturbances and commodities. This iteration on supports
a single mine_type. The reason for this is that these changes should be
merged into the base branch soon and built upon as we move to supporting
multiple mine types and disturbances.

The backend handles expiring updated records. They are expired when an
update is sent to an existing mine_type record with a new
mine_tenure_type_code value. This retains a history of the changes while
always having knowledge of the current "active" record.

The frontend handles making the decision about which endpoint to call.
Most notably, when a user is "updating" a mine record from having no
mine types to having a mine type, this information must be POSTed to the
mines/mine_types endpoint, despite that it is done in the context of a
mine record "update."

* Support disturbance type deletion

* Add model validation for tenure type & disturbance

Some disturbances are invalid for certain tenure types. This vaidation
enforces those rules.

* Rename migration file to fix build