MDS-455 Feature/Add commodity types and fully support Mine Type management (#311)
* Update Mine Tenure Type to follow standard pattern

* Use codes instead of ID
* Use description instead of name

* Update Mine Tenure Type name and resource

* Add tenure type relation to disturbance response

* Standardize quote type

This was also to trigger a rebuild.

* Add POST /mines/mine_types endpoint

In accordance with our RESTful architecture, there must exist an
endpoint for creating a mine, and then a separate endpoint for creating
a mine type associated with that mine.

This pattern will then extend to disturbance type, which will be
associated with a mine type.

* Add POST /mines/mine_types/details endpoint

This endpoint accepts mine disturbance data and allows for the creation
of disturbance records that are associated with a mine type.

* Add mine_type_detail to /mines/guid response

* Add PUT /mines/mine_types and frontend logic

The frontend now makes separate requests when updating a mine record:
one for the mine_identity/details and one for mine_type. This will be
built-upon with disturbances and commodities. This iteration on supports
a single mine_type. The reason for this is that these changes should be
merged into the base branch soon and built upon as we move to supporting
multiple mine types and disturbances.

The backend handles expiring updated records. They are expired when an
update is sent to an existing mine_type record with a new
mine_tenure_type_code value. This retains a history of the changes while
always having knowledge of the current "active" record.

The frontend handles making the decision about which endpoint to call.
Most notably, when a user is "updating" a mine record from having no
mine types to having a mine type, this information must be POSTed to the
mines/mine_types endpoint, despite that it is done in the context of a
mine record "update."

* Support disturbance type deletion

* Add model validation for tenure type & disturbance

Some disturbances are invalid for certain tenure types. This vaidation
enforces those rules.

* Add /mines/commodity_codes endpoint

This is an informational endpoint that returns all of the valid
commodity codes.

* Resolve migration name conflict

* Refactor serialized JSON to remove extra logic

* Add support for adding/removing commodity records

* DRY up the POST method

* Add support and validation for deleting mine types

* Update commodities JSON and default value fetching

* Convert urls to kebab case

* Remove active_ind from respnse & filter inactive

* Fix create_data error on mine_type insertion

* Fix incorrect table comment