MDS-1438 Feature/ Add 0, 1, or n roles to a new party (#603)
* Fix linting errors

* Roles WIP step 2

* Add/Remove role items in UI

* Create roles

* Add simple relationship types

* Search for mines with autocomplete

* Some style updates

More style changes to come, and mine search validation still needs to be
fixed. These are just initial changes based on UI/UX feedback.

* More style changes

* Update styles, validation, and UX

* Update tests

* Add new snapshot

* Remove date validation & improve UX

New Role panels are opened automatically

* Disable 'Next' button for invalid payload

This addresses a runtime error that could occur when some fields were
missing. There is likely a deeper issue that is OOS for the current
task, but this solution hides it from the end user.

Also updated the snapshots from latest develop.

* Handle org type for validation

I've removed the submission that triggers the form validation because
it was causing a runtime exception that crashed the app. Instead, the
Next/Submit button is disabled, and dirty fields have validation
applied.

This can be iterated on to improve user experience if needed.