Refactor to make the implementation less terrible
There needs to be some more investigation into using
componentWillReceivePropsjand/or this.forceUpdate().

This continues to have a few issues:
- This solution still lives in the component (meaning it must be
implemented in every component with a download option individually)
- The side-effecting method that fetches the download link is triggered
on every render. The early-exit means it won't make the API request once
the value is stored in state, but the async nature of setState means
that the request can still be made multiple times