Check store for static content - mine status only
This establishes a potential pattern for avoiding repeated API requests
for static content. If the content exists in the store, that value will
be returned by axios instead of making an API request.

This removes the need for managing any state around whether or not the
static content has been fetched, handling individual API request
failures, etc.