Update App/index imports/exports to fix deps tree
The dependency tree was being incorrectly loaded due to index.js having
exports. This became an issue as soon as one of those exports were used
(in the case of CustomAxios importing the store).

This updates the dependency tree such that Jest will not attempt to
mount on a non-existent root DOM element.