Support refreshing any data type
Next steps:
1) Remove the `records` envelope. To do this, the store actions will
need to stop extracting from the `records` property. The code in the
Axios chain(s) will need to handle unwrapping the data. The benefit here
is that the data can then be stored as-is as either an object or as an
array. This makes the type of envelope provided by the API addressable
on a case-by-case basis.

2) Performance improvements to first-load. Leveraging the changes made
to the store as a caching mechanism for static content, the heavy first
load can be decreased. Any component requiring static assets can request
them directly.