Use Redux Store as staticContent caching mechanism & add refresh button (#865)
* Check store for static content - mine status only

This establishes a potential pattern for avoiding repeated API requests
for static content. If the content exists in the store, that value will
be returned by axios instead of making an API request.

This removes the need for managing any state around whether or not the
static content has been fetched, handling individual API request
failures, etc.

* Update App/index imports/exports to fix deps tree

The dependency tree was being incorrectly loaded due to index.js having
exports. This became an issue as soon as one of those exports were used
(in the case of CustomAxios importing the store).

This updates the dependency tree such that Jest will not attempt to
mount on a non-existent root DOM element.

* Standardize selectors/reducers with API stubbing

* Update tests

* Remove unneceesary line in test

* Remove all references to tracking optionsLoaded

* Remove all unnecessary console.logs

* Wrap application codes in records envelope

* Add RefreshButton component

* Add RefreshButton to mine view for Variances

This is a POC for passing actions/requests into a child component and
using the refresh functionality for updating any list response (not just
static content values).

* Add default values to facilitate non-list values

* Add permits to mine view refresh

* Fix Makefile flags bug

* Fix inconsistent button height

* Support refreshing any data type

Next steps:
1) Remove the `records` envelope. To do this, the store actions will
need to stop extracting from the `records` property. The code in the
Axios chain(s) will need to handle unwrapping the data. The benefit here
is that the data can then be stored as-is as either an object or as an
array. This makes the type of envelope provided by the API addressable
on a case-by-case basis.

2) Performance improvements to first-load. Leveraging the changes made
to the store as a caching mechanism for static content, the heavy first
load can be decreased. Any component requiring static assets can request
them directly.