Remove all NOW prefixes & update endpoint namespace
The now-submissions namespace will be used to keep more typical
namespaces freed up for future API iterations, without needed to break
the existing API. The idea being that the data shape returned from  this
endpoint may be temporary, so we may deprecate the endpoint or remove it
entirely if no systems are consuming it.

The primary area of concern is the column names, which are inconsistent
with the other MDS columns. In an effort to avoid attempting to predict
the future, we are namespacing this API with the column names being
marshalled as-is.