Fix file upload crash with WDS
It looks like the setup for Filepond changed as we updated to a new
version. The result was that we were rendering `undefined`, which was
crashing WDS. I removed the code causing the problem.

After removing it, everything behaved as-normal. I haven't found any
regressions or missing functionality that this state-based code was
intended to support (at least in the current version of Filepond).