Fix crash when switching between list/map tabs
I believe the cause of the crash was that the updated loading logic
didn't force the currently-rendered component (list or map) to unmount
prior to the Redux state change. This meant that:

1. Either the new tab needed to be rendered before the data was updated
(causing a crash on the new tab) OR
2. The old tab would still be rendered as the data was updated (causing
a crash in the old tab).

This is somewhat speculative, based on the behaviour of the loading
wrapper. By checking the loaded state in the child component and
conditionally-rendering there, the mount-unmount problem seems to be
resolved.