MDS-2669: MineSpace and Core CRR Create/Edit/View Issues (#1231)
Main
MDS-2669: "Editing a CRR on MineSpace does not show existing documents, and adding new documents overwrites the set (and it only will upload a single document if multiple are provided)": Fixes a variety of issues with CRR on MineSpace:

Don't allow proponents to delete documents added to a report in the Edit view!
Fixes an issue where adding documents to an existing CRR in MineSpace did not work. We were improperly packaging the payload sent to the PUT method.
Fixes an issue with displaying the documents for each CRR record in the MineSpace table
MDS-2668: "Uploading CRRs causes the drag & drop box to move, causing users to miss the upload box on subsequent uploads": Now the uploaded files box is capped at 130px in the Add CRR view (see pic below) and 260px (the same as the Filepond upload box height) in the Edit CRR view. positions of the file upload/uploaded files components have been swapped to solve this issue (thanks Ryan for the better idea!):
image

Other
Fixes various "each child must have a key prop" error in Core and MS
Fixes the styling of the "remove" buttons in the MineSpace "uploaded files" component to use the proper Ant Design styling:
image
Various other small tweaks to tidy/remove redundant code