MDS-2929: Implement Bond Closed Date/Notes (Modify Bond Confiscate/Release Flow) (#1383)
Main
Modifies the bond release/confiscation flow by allowing users to set the release/confiscation date as well as provide release/confiscation notes. Note: Adding release/confiscation documents is supported in the Edit Bond modal after the bond has been closed. This was decided because it would add a lot more scope if we allowed the user to select/upload multiple release/confiscation document types in the modal.
Adds the columns closed_date and closed_note to the tables bond and bond_history and updates the bond_view to account for these new columns. Note: Securities reporting will be addressed in another ticket to account for these new columns.
Updates the View Bond Modal by displaying the closed date/notes if the bond is closed.
Updates the Edit Bond Modal by allowing the editing of the closed date/notes if the bond is closed. Note: If the bond is confiscated, the bond type cannot be edited, as changing the bond type to cash when the bond was confiscated was an enforced business rule.
Note: The language for the closed date/notes changes dynamically depending on whether the bond was released or confiscated.
Note: Existing bond and bond history entries, if they are released or confiscated, have their closed date set to the update timestamp.

Screenshots
image
image
image
image

Other
Moves the Notes field in the Edit Bond Modal up to the other general information area rather than the Institution area