Fix Loading Bar Issues (#1386)
I have a PR for the react-redux-loading-bar repo but in the meantime, we can import the source for the LoadingBar and related classes into our project. When and if the PR gets in we can resort to using the package for the modal wrapper's loading bar (the navbar still uses the package). I resolved the issue by using an alternate animation approach (using width as opposed to translate3d). I also fixed an issue where the width being set in the navbar's loading bar which was preventing it from functioning.