MDS-3428: Add Notice of Work Form Automatically to FAP (#1661)
Main
When a Notice of Work is imported or every time Technical Review is completed, a Notice of Work Form document is automatically generated and included in the final application package. All previous Notice of Work Form documents that were in the final application package are removed from it.
Renames the document Technical Review to Notice of Work Form.
Notice of Work Form documents are no longer excluded from the NoW documents lists.
image
Other
Sets a "submitting" state on the "Update NoW Progress" modal so the buttons cannot be spammed while submitting.