MDS-3652: PDF Viewer Implementation (#1725)
Main
Implements a Document Viewer component to display documents within the application.
Currently, this is implemented for PDFs and uses Syncfusion's PDFViewer component. In the future, we can use Syncfusion's other components for displaying "word" and "sheet" documents in the app as well.
All documents that are openable by the Document Viewer are opened instead of being downloaded. This is the case for all documents that exist within Core, and not for NoW submission documents that have not yet been imported, for example.
Screenshots
image

Other
Creates and implements a Document Link component for consistently displaying document links.
Tidying of various code (e.g., reduce code duplication in actionlessNetworkCalls.js.