[MDS-4153] -  vfcbc field update (#2635)
* Update form elements and rename file extension

Rearranged several form components on the application review page to offer a more intuitive user workflow. The 'term_of_application' field was moved and a new field 'mine_purpose' was added to offer more clarity in the form. The 'RenderRadioButtons' component file was renamed with '.tsx' extension to adhere to the TypeScript naming convention. Additionally, the 'RenderRadioButtons' component was refactored to use correct TypeScript types to improve code readability and maintainability.

* Update value of radio button options in ReviewNOWApplication

Values of radio button options for 'Exploration Activities' and 'Production Mining' in the ReviewNOWApplication form were misaligned. These values were updated to "Exploration Activities" and "Production Mining" respectively.

* Update ReviewNOWApplication.spec.js.snap

* Update field validation

* Update field validation

* updated test snaps