Add replace document functionality
Implemented document replace functionality that allows user to replace an existing document with a new one. Replace feature is available in both minespace and core web applications. Made necessary changes in UI to handle document replace scenarios and update document details accordingly. Added file type validation to ensure the new document is of the same type as the original. Refactored the Document model to maintain all versions of a document, instead of excluding the latest version. Added a new endpoint in document manager API to handle the replace document request and updated the tusd hooks to handle the post finish event correctly. Updated document list table in minespace and core web to include 'replace' option in action menu.