"Update number of previous versions calculation in document test
The calculation for the expected number of previous versions in document.spec.ts was amended. Previously, it was subtracting 1 from the length of the versions array, which does not correctly reflect the total number of versions. The subtraction was removed to ensure accurate calculation of the number of previous versions."