[MDS-5448] Refactor file upload process for better error handling (#2664)
* Refactor file upload process for better error handling

Revised the file upload process to poll for upload status instead of simply waiting for the file to upload or fail. This revision is designed to provide more descriptive errors when file uploads fail since it can now catch errors that occur mid-upload. Also added test cases for the new behaviour, and the necessary database migrations to store upload status information.

* match common folder

* updated snapshots

* update docman db migration to set previous document status' to "Success"

---------