[MDS-5470] - Optimized CSV export of Notice of Work details (#2671)
* Optimized CSV export of Notice of Work details

The previous implementation was caching the entire output CSV. This was replaced with a generator to stream the CSV content to the client as it's being produced, making the system more efficient when large data sets are processed. This change will improve performance and memory usage. Plus, a minor change was made in the description string to correct a spelling error.

* Remove cache.clear() and lowered yield_per