[MDS-5353] ESUP Modal updates (#2705)
* Add Explosives Permit form and Remove token after complete download

New form has been created to handle the addition of explosives permits. Form will handle business rules validations, state, and UI. Ant.design and redux-form have been used to handle state management and UI.

In addition, a minor bug fix was implemented in the document.py file to ensure that the cache for downloading documents get deleted after the download operation completes. This is to prevent reuse of the same token for downloading multiple documents, which could be a potential security vulnerability.

Styles have also been tweaked for MinePermitTable, Modal, and new styles for ExplosivesPermits have been created.

Re-ordered import statements to keep maintain consistency and readability across the code base.

* "Remove unnecessary defaultProps from DocumentTable component

The defaultProps object in DocumentTable.tsx was removed because it was not being used bring any value to the declared component. The methods: openModal and closeModal, defined in this object, were not applied in the component. It enhances readability by removing unused parts of code."

* Remove unused dataIndex from ExplosivesPermitTable

* address PR comments