Add history view and comparison feature for Explosives Permits
Added a feature to show the history of amendments in an Explosives Permit. This feature provides a clear way to track changes and the reasons for the modifications in the permit overtime. Updates were made in respective models, interfaces and file structures. Code enhancement also includes providing table views and detailed comparisons of permits history for each amendment. This assists users to see differences between various versions of permit and understand the changes that were made. User interface was also improved to allow better navigation.