Refactor MagazineFormNew from functional to class component
The MagazineFormNew component in the core-web components has been refactored from a functional to a class-based React component. The functional components were changed to class methods to behave as part of the class instance. PropTypes were replaced with TypeScript style prop typing and React.useState was replaced with traditional this.state usage. This refactoring was done to align with the code base that mostly use class-based components, improve the readability of the code, and better manage the component level state.