Added an optional trigger to generate documents
The code was updated to allow optional generation of documents. Previously, every update to the permit or amendment would trigger regeneration of related documents, which was not always necessary. Now, apply a new parameter 'generate_documents' to control whether to generate documents in the function of updating explosives permit or amendment.