[MDS-5662]Update report selection and submission flow (#2893)
* Update report selection and submission flow

The code incorporates multiple improvements for the report selection and submission flow. It adds 'is_common' field to report definitions to categorize commonly used reports. The compliance code is formatted for report names to provide easy readability. The 'Report Code Requirement' field was made into a dropdown selection for ease of use, replacing the manual inputs.

* update snaps

* merge duplicate useEffect dependencies

* update report interface location