Add region to project summary and create regions table
A new 'regions' table is created with a primary key 'regional_district_id' and other requisite fields such as 'name', 'create_user' etc. The 'project_summary' table is updated with a new field 'regional_district_id' and a foreign key constraint is established to link it with the 'regions' table.